+++ Symlink/functions/Set-Symlink.ps1
@@ -18,10 +18,11 @@
 	The new value for the property to take.
 	
 .PARAMETER WhatIf
-	wip
+	Prints what actions would have been done in a proper run, but doesn't
+	perform any of them.
 	
 .PARAMETER Confirm
-	wip
+	Prompts for user input for every "altering"/changing action.
 	
 .INPUTS
 	Symlink[]
@@ -42,12 +43,21 @@
 	This command will change the name of the symlink called "data", to the new
 	name of "WORK". From now on, there is no symlink named "data" anymore.
 	
+.EXAMPLE
+	PS C:\> Set-Symlink -Name "data" -Property "Path" -Value "~\Desktop\Files"
+	
+	This command will change the path of the symlink called "data", to the new
+	location on the desktop. The old symbolic-link item from the original
+	location will be deleted, and the a new symbolic-link item will be created
+	at this new location.
+	
 #>
 function Set-Symlink {
 	
 	[CmdletBinding(SupportsShouldProcess = $true)]
 	param (
 		
+		# Tab completion.
 		[Parameter(Position = 0, Mandatory = $true, ValueFromPipelineByPropertyName)]
 		[string]
 		$Name,
@@ -63,10 +73,10 @@ function Set-Symlink {
 	)
 	
 	process {
-		Write-Verbose "Processing the symlink: '$Name'."
 		# Read in the existing symlinks.
 		$linkList = Read-Symlinks
 		
+		Write-Verbose "Changing the symlink: '$Name'."
 		# If the link doesn't exist, warn the user.
 		$existingLink = $linkList | Where-Object { $_.Name -eq $Name }
 		if ($null -eq $existingLink) {
@@ -79,52 +89,59 @@ function Set-Symlink {
 			Write-Verbose "Changing the name to: '$Value'."
 			
 			# Validate that the new name is valid.
-			if ([system.string]::IsNullOrWhiteSpace($Name)) {
+			if ([System.String]::IsNullOrWhiteSpace($Name)) {
 				Write-Error "The name cannot be blank or empty!"
 				return
 			}
 			# Validate that the new name isn't already taken.
 			$clashLink = $linkList | Where-Object { $_.Name -eq $Value }
 			if ($null -ne $clashLink) {
-				Write-Error "The name: '$Value' is already taken."
+				Write-Error "The name: '$Value' is already taken!"
 				return
 			}
 			
 			$existingLink.Name = $Value
-			
-		}elseif ($Property -eq "Path") {
-			Write-Verbose "Changing the path to: '$Path'."
+		}
+		elseif ($Property -eq "Path") {
+			Write-Verbose "Changing the path to: '$Value'."
 			# First delete the symlink at the original path.
-			$existingLink.DeleteFile()
+			if ($PSCmdlet.ShouldProcess($existingLink.FullPath(), "Delete Symbolic-Link")) {
+				$existingLink.DeleteFile()
+			}
 			
 			# Then change the path property, and re-create the symlink
 			# at the new location.
 			$existingLink._Path = $Value
-			$existingLink.CreateFile()
-			
-		}elseif ($Property -eq "Target") {
+			if ($PSCmdlet.ShouldProcess($existingLink.FullPath(), "Create Symbolic-Link")) {
+				$existingLink.CreateFile()
+			}
+		}
+		elseif ($Property -eq "Target") {
 			Write-Verbose "Changing the target to: '$Value'."
 			
 			# Validate that the target exists.
-			if ((Test-Path -Path ([System.Environment]::ExpandEnvironmentVariables($Value))) -eq $false) {
+			if (-not (Test-Path -Path ([System.Environment]::ExpandEnvironmentVariables($Value)) `
+					-ErrorAction Ignore)) {
 				Write-Error "The target path: '$Value' points to an invalid location!"
 				return
 			}
 			
 			# Change the target property, and edit the existing symlink (re-create).
 			$existingLink._Target = $Value
-			$existingLink.CreateFile()
-			
-		}elseif ($Property -eq "CreationCondition") {
+			if ($PSCmdlet.ShouldProcess($existingLink.FullPath(), "Update Symbolic-Link target")) {
+				$existingLink.CreateFile()
+			}
+		}
+		elseif ($Property -eq "CreationCondition") {
 			Write-Verbose "Changing the creation condition."
-			$existingLink._Condition = $Value
 			
+			$existingLink._Condition = $Value
 			# TODO: Operate if condition result is different from previous state.
 		}
 		
 		# Re-export the list.
-		Write-Verbose "Re-exporting the modified database."
-		Export-Clixml -Path $script:DataPath -InputObject $linkList | Out-Null
+		if ($PSCmdlet.ShouldProcess("$script:DataPath", "Overwrite database with modified one")) {
+			Export-Clixml -Path $script:DataPath -InputObject $linkList -WhatIf:$false -Confirm:$false | Out-Null
+		}
 	}
-	
 }
